Set up trades (but only the bank)
This commit sets up trades with other players but makes a distinction
for the bank (i.e. "maritime trade" as it is called in the rules), which
is the only trade that is implemented.

In addition, a class has been set up to deal with representing
collections of fungible cards.