Implement placing settlements
This commit re-arranges things so that the AuthenticatedPlayer class
knows about the piece classes, which no longer know about players but do
know what color they are.

Also the players have piece pools to allow only a finite number of
pieces to be played.

Known bugs:
1) implementation failure: settlements breaking roads of other players
is explicitly mentioned as allowed in the manual, so that validation is
wrong
2) sea tiles being represented as undefined means that a road cannot be
placed on a coast if only a road on the coast of a different hex is
providing the necessary connection, because the algorithm would need the
missing sea tile to be able to confirm that the target edge is
connected.