Prevent roads going through the settlements of other players
This commit chagnes the logic to implement the correct rule that a
settlement belonging to another player blocks the roads of other players
from going past it.

It also tidies up the BoardVisualization and PlayerVisualization classes
to use better dependency injection and removes the lazy use of
inheritance.

Furthermore, the documentation has been greatly expanded.